Implement ARP / SYN scanner
This is a much faster and better implementation of a pure go scanner.
It uses gopacket to perform arp scanning to find devices on the network
along with their mac addresses. It then performs a SYN scan on those
found devices to detect when port 22 is open. Caveats are that this
requires ops to be run with root privileges and has a dependency on
libpcap.