Remove use of sqlite db, server service, and ansible dependency
The sqlite DB and server service are not necessary. There's no need to
store servers in a database as we are actively scanning the network
periodically and providing updates via events. Configs are now stored in
a json file.

Similarly, having a dependency on ansible just to get hostname and
os info seems silly. The same is now achieved just using plain 'ol
ssh and "uname -a" command.